Move internal object classes to the already existing internal dir inthe sync dir
