+++ Sync/SyncDataExchange/Internal/Object/ObjectInterface.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * @license     GNU/GPLv3 http://www.gnu.org/licenses/gpl-3.0.html
  */
 
-namespace MauticPlugin\IntegrationsBundle\Internal\Object;
+namespace MauticPlugin\IntegrationsBundle\Sync\SyncDataExchange\Internal\Object;
 
 interface ObjectInterface
 {
