+++ Tests/Unit/Sync/DAO/InputOptionsDAOTest.php
@@ -16,9 +16,9 @@ namespace MauticPlugin\IntegrationsBundle\Tests\Unit\Sync\DAO;
 use DateTimeImmutable;
 use DateTimeZone;
 use MauticPlugin\IntegrationsBundle\Exception\InvalidValueException;
-use MauticPlugin\IntegrationsBundle\Internal\Object\Contact;
 use MauticPlugin\IntegrationsBundle\Sync\DAO\Sync\InputOptionsDAO;
 use MauticPlugin\IntegrationsBundle\Sync\DAO\Sync\ObjectIdsDAO;
+use MauticPlugin\IntegrationsBundle\Sync\SyncDataExchange\Internal\Object\Contact;
 use MauticPlugin\IntegrationsBundle\Sync\SyncDataExchange\MauticSyncDataExchange;
 
 class InputOptionsDAOTest extends \PHPUnit_Framework_TestCase
